merge

+++ database/migrations/2026_02_23_175430_create_expenses_table.php
@@ -17,7 +17,7 @@ return new class extends Migration
             $table->foreignId('payer_id')->constrained('users') ->onDelete('cascade') ;
             $table->foreignId('category_id')->constrained('categories')->onDelete('cascade') ; 
             $table->string('title') ;
-            $table->float('amount') ;
+            $table->int('amount') ;
 
             $table->timestamps();
         });
